Fixing DSA sample

+++ tests/gl-430-draw-vertex-attrib-binding.cpp
@@ -246,7 +246,7 @@ private:
 			glUnmapBuffer(GL_UNIFORM_BUFFER);
 		}
 
-		glViewportIndexedf(0, 0, 0, GLfloat(WindowSize.x), GLfloat(WindowSize.y));
+		glViewportIndexedf(0, 0, 0, WindowSize.x, WindowSize.y);
 
 		glClearBufferfv(GL_COLOR, 0, &glm::vec4(1.0f, 0.5f, 0.0f, 1.0f)[0]);
